Fix the awful view hack
footy/views.py had an awful hack in that enumerated every match when loading the /matches page.
With the import of the new dataset, this makes the load time for the match page ridiculously long.
The only reason for the hack was that the match_list template took TMS's as an iterator rather
than MS's. A quick (though not completely correct) fix duplicates the match_list template with one
that also takes MS's. Also this commit cuts down the ridiculous number of paginator blocks at the
end of matches page, again, not in the completely correct way.